예외 처리및 API 작성 완료:

+++ controllers/PostController.php
@@ -2,7 +2,7 @@
 require 'function.php';
 
 const JWT_SECRET_KEY = "TEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEYTEST_KEY";
-
+$content_regex="/^[\p{Z}\s]*(?:[^\p{Z}\s][\p{Z}\s]*){2,300}$/u";
 $res = (Object)Array();
 header('Content-Type: json');
 $req = json_decode(file_get_contents("php://input"));
@@ -25,7 +25,23 @@ try {
 
         case "createPost":
             http_response_code(200);
-            $res->result = createPost($req->user_id,$req->content,$req->post_image);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+            if(preg_match($content_regex, $req->content)!=true){
+                failRes($res, "올바르지 않은 내용 입니다.", 211);
+                break;
+            }
+
+            if(isset($req->post_image) and !checkImageExt($req->post_image)){
+                failRes($res,"이미지 파일 형식 오류", 210);
+                break;
+            }
+            $res->result = createPost($userId,$req->content,$req->post_image);
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "게시글 작성 성공";
@@ -33,6 +49,24 @@ try {
             break;
         case "updatePost":
             http_response_code(200);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+
+
+            if(preg_match($content_regex, $req->content)!=true){
+                failRes($res, "올바르지 않은 내용 입니다.", 211);
+                break;
+            }
+
+            if(!checkPostPermission($userId,$req->post_id)){
+                failRes($res, "게시글 수정 권한이 없습니다.", 211);
+                break;
+            }
             $res->result = updatePost($req->post_id,$req->content,$req->post_image);
             $res->isSuccess = TRUE;
             $res->code = 100;
@@ -42,7 +76,10 @@ try {
         case "getPost":
             http_response_code(200);
             //user정보 받아서 좋아요 표시해야함
-            $res->result = getPost($vars["post-id"]);
+            if(!$res->result = getPost($vars["post-id"])){
+                failRes($res, "존재하지않는 게시글입니다..", 211);;
+                break;
+            }
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "게시글 조회 성공";
@@ -51,7 +88,13 @@ try {
         case "getPostList":
             http_response_code(200);
             //user정보 받아서 좋아요 표시해야함
-            $res->result = getPostList($_GET["filter"]);
+            $keyword = preg_replace("/[ #\&\+\-%@=\/\\\:;,\.'\"\^`~\_|\!\?\*$#<>()\[\]\{\}]/i", "", $_GET["keyword"]);
+            $res->result = getPostList($keyword,$_GET["lastIdx"]);
+            if(empty($res->result)){
+                failRes($res, "더 이상 게시글이 없습니다", 211);;
+                break;
+            }
+            
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "게시글 목록 조회 성공";
@@ -59,6 +102,18 @@ try {
             break;
         case "deletePost":
             http_response_code(200);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+
+            if(!checkPostPermission($userId,$_GET["post-id"])){
+                failRes($res, "게시글 수정 권한이 없습니다.", 211);
+                break;
+            }
             $res->result = deletePost($_GET["post-id"]);
             $res->isSuccess = TRUE;
             $res->code = 100;
@@ -67,13 +122,20 @@ try {
             break;
         case "likePost":
             http_response_code(200);
-            if(!checkIfExist("PostLike",["user_id","post_id"],[$req->user_id,$req->post_id])) {
-                $res->result = likePost($req->user_id, $req->post_id);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+            if(!checkIfExist("PostLike",["user_id","post_id"],[$userId,$req->post_id])) {
+                $res->result = likePost($userId, $req->post_id);
                 $res->isSuccess = TRUE;
                 $res->code = 100;
                 $res->message = "게시글 좋아요 성공";
             }else{
-                $res->result = deleteLikePost($req->user_id, $req->post_id);
+                $res->result = deleteLikePost($userId, $req->post_id);
                 $res->isSuccess = TRUE;
                 $res->code = 100;
                 $res->message = "게시글 좋아요 취소 성공";
@@ -82,7 +144,19 @@ try {
             break;
         case "createComment":
             http_response_code(200);
-            $res->result = createComment($req->post_id,$req->user_id,$req->content);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+            if(preg_match($content_regex, $req->content)!=true){
+                failRes($res, "올바르지 않은 내용 입니다.", 211);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+            $res->result = createComment($req->post_id,$userId,$req->content);
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "댓글 생성 성공";
@@ -91,6 +165,23 @@ try {
         case "updateComment":
             http_response_code(200);
             //user 인증
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+
+            if(!checkCommentPermission($userId,$req->comment_id)) {
+                failRes($res, "댓글 수정 권한이 없습니다.", 211);
+                break;
+            }
+
+            if(preg_match($content_regex, $req->content)!=true){
+                failRes($res, "올바르지 않은 내용 입니다.", 211);
+                break;
+            }
             $res->result = updateComment($req->comment_id,$req->content);
             $res->isSuccess = TRUE;
             $res->code = 100;
@@ -100,6 +191,18 @@ try {
         case "deleteComment":
             http_response_code(200);
             //user 인증
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            if (!isValidHeader($jwt, JWT_SECRET_KEY)) {
+                failRes($res,"유효하지 않은 토큰입니다",201);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+            $userId = getUserNoFromHeader($jwt, JWT_SECRET_KEY);
+
+            if(!checkCommentPermission($userId,$_GET["comment-id"])) {
+                failRes($res, "댓글 수정 권한이 없습니다.", 211);
+                break;
+            }
             $res->result = deleteComment($_GET["comment-id"]);
             $res->isSuccess = TRUE;
             $res->code = 100;
